Switched to Python 3.5
* README.md:

The documentation now states that the toolbox is written for Python 3.5

* environment.yml: Changed requested Python version to 3.5

* scripts/01_03_Matrizenrechnung: Use built-in matrix multiplication.

* scripts/03_01_Einlesen_einer_ASCII_Datei.py,
  scripts/03_02_Einlesen_mehrerer_ASCII_Dateien.py:

Use byte-string while reading file.

* scripts/04_07_Fouriertransformation.py: Removed implicit integer division.

+++ scripts/04_07_Fouriertransformation.py
@@ -37,8 +37,8 @@ Y = np.fft.fft(y)/n  # FFT Berechnung und Normierung
 
 # Abschneiden der zweiten Vektorhälfte um einseitge FFT
 # für Zeitreihen mit realen Werten zu erhalten
-Y = 2 * Y[range(n/2)] # Amplitude der Fouriertransformierten
-freq = freq[range(n/2)] # Einseitiger Frequenzbereich
+Y = 2 * Y[range(int(n/2))] # Amplitude der Fouriertransformierten
+freq = freq[range(int(n/2))] # Einseitiger Frequenzbereich
 
 # Plot der Fouriertransformation
 axes[1].plot(freq, abs(Y), 'r')
